Completed Exercises 4.18-4.19: bloglist expansion, step6 and 7

+++ part4/bloglist-backend/controllers/blogsRouter.js
@@ -1,23 +1,42 @@
+const { getDecodedToken } = require('../utils/token')
+
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-const userPopulateObj = { path: 'user', select: '-password -blogs' }
+
+const userPopulateObj = { path: 'user', select: '-passwordHash -blogs' }
+
 
 blogRouter.get('/', async (request, response) => {
+
   const blogs = await Blog.find({}).populate(userPopulateObj)
+
   response.json(blogs)
 })
 
 blogRouter.post('/', async (request, response) => {
 
-  // get userId
-  const user = await User.findOne({})
+  // get token
+  const decodedToken = getDecodedToken(request)
+  if (!decodedToken || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+
+  // get user
+  const user = await User.findById(decodedToken.id)
+  if(!user){
+    return response.status(401).json({
+      error: 'user was not found'
+    })
+  }
 
   // Save blog to db
   const newBlog = await (new Blog({...request.body, user: user.id})).save()
 
   await User.findOneAndUpdate(user.id, { blogs: user.blogs.concat(newBlog.id) })
-  response.status(201).json(newBlog)
+  return response.status(201).json(newBlog)
 })
 
 // ----------------------------------------
@@ -38,6 +57,14 @@ blogRouter.put('/:id', async (request, response) => {
   const id = request.params.id
   if(!id) return response.status(400).json({ error: 'no id' })
 
+  // get token
+  const decodedToken = getDecodedToken(request)
+  if (!decodedToken || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+
   const blogUpdateData = { ...request.body }
 
   const retBlog = await Blog.findByIdAndUpdate(id, blogUpdateData, { new: true })
@@ -53,7 +80,25 @@ blogRouter.delete('/:id', async (request, response) => {
   const id = request.params.id
   if(!id) return response.status(400).json({ error: 'no id' })
 
+  // get token
+  const decodedToken = getDecodedToken(request)
+  if (!decodedToken || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+
+  // get user
+  const user = await User.findById(decodedToken.id)
+  if(!user){
+    return response.status(401).json({
+      error: 'user was not found'
+    })
+  }
+
   const ret = await Blog.findByIdAndRemove(id)
+
+  await User.findOneAndUpdate(user.id, { blogs: user.blogs.filter(blog => blog.id.toString() !== id) })
 
   if(ret){
     return response.status(204).end()
